+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.6.0
+### Added
+- Documentation everywhere.
+
+### Changed
+- Make compatible with NQE 0.5.
+- Use supervisor only in peer manager.
+- API quality of life changes.
+- Exposed module is now only `Haskoin.Node`.
+
+### Removed
+- No more direct access to internals.
+
 ## 0.5.2
 ### Changed
 - Improve dependency definitions.